Relocated "member-type-root" folder
