+++ src/Umbraco.Web.UI.Client/src/packages/members/member-type/index.ts
@@ -2,8 +2,10 @@ import './components/index.js';
 
 export * from './components/index.js';
 export * from './constants.js';
+export * from './member-type-root/index.js';
 export * from './modal/member-type-picker-modal.token.js';
 export * from './repository/index.js';
 export * from './tree/index.js';
 export * from './workspace/constants.js';
+
 export type * from './types.js';
