+++ src/Umbraco.Web.UI.Client/src/packages/members/member-type/workspace/constants.ts
@@ -1,2 +1,2 @@
-export * from './member-type/constants.js';
-export * from './member-type-root/constants.js';
+export * from './member-type-workspace.context-token.js';
+export const UMB_MEMBER_TYPE_WORKSPACE_ALIAS = 'Umb.Workspace.MemberType';
