+++ src/Umbraco.Web.UI.Client/src/packages/members/member-type/workspace/index.ts
@@ -1,2 +1,2 @@
-export * from './member-type/index.js';
-export * from './member-type-root/index.js';
+export * from './member-type-workspace.context-token.js';
+export * from './constants.js';
