+++ src/Umbraco.Web.UI.Client/src/packages/members/member-type/workspace/member-type-workspace.context.ts
@@ -1,5 +1,5 @@
-import type { UmbMemberTypeDetailModel } from '../../types.js';
-import { UMB_MEMBER_TYPE_DETAIL_REPOSITORY_ALIAS, UMB_MEMBER_TYPE_ENTITY_TYPE } from '../../index.js';
+import type { UmbMemberTypeDetailModel } from '../types.js';
+import { UMB_MEMBER_TYPE_DETAIL_REPOSITORY_ALIAS, UMB_MEMBER_TYPE_ENTITY_TYPE } from '../index.js';
 import { UmbMemberTypeWorkspaceEditorElement } from './member-type-workspace-editor.element.js';
 import { UMB_MEMBER_TYPE_WORKSPACE_ALIAS } from './constants.js';
 import {
