--- src/Umbraco.Web.UI.Client/src/packages/members/member-type/workspace/member-type/index.ts
@@ -1,2 +1,0 @@
-export * from './member-type-workspace.context-token.js';
-export * from './constants.js';
